+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-[3.1.X] - 2023-xx-xx
+[3.2.0] - 2023-xx-xx
 --------------------
 * New Features
   * Added tests for warnings, logging messages, and errors in the Instrument
@@ -25,6 +25,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Updated docstring references to `pysat.utils.files` in other modules.
   * Remove Sphinx cap
   * Add pandas cap
+  * Remove deprecated `pysat_testing2d` instrument
+  * Remove deprecated meta children info
 
 [3.1.0] - 2023-05-31
 --------------------